+++ Source/AwsCommonRuntimeKit/io/retryer/CRTAWSRetryToken.swift
@@ -7,13 +7,12 @@ public final class CRTAWSRetryToken {
     var rawValue: UnsafeMutablePointer<aws_retry_token>
 
     public init(retryStrategy: CRTRetryStrategy, allocator: Allocator = defaultAllocator) {
-        let intPointer: UnsafeMutableRawPointer = fromPointer(ptr: 1)
-        let atomicVar = aws_atomic_var(value: intPointer)
+        let atomicVar = Atomic<Int>(1)
         let retryStrategyPtr: UnsafeMutablePointer<CRTRetryStrategy> = fromPointer(ptr: retryStrategy)
         let retryToken = aws_retry_token(allocator: allocator.rawValue,
-                                        retry_strategy: retryStrategy.rawValue,
-                                        ref_count: atomicVar,
-                                        impl: retryStrategyPtr)
+                                         retry_strategy: retryStrategy.rawValue,
+                                         ref_count: atomicVar.rawValue,
+                                         impl: retryStrategyPtr)
         let retryTokenPtr: UnsafeMutablePointer<aws_retry_token> = fromPointer(ptr: retryToken)
         self.rawValue = retryTokenPtr
     }
